# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  flood: 'flood',
  cyclone: 'cyclone',
  earthquake: 'earthquake',
  wildfire: 'wildfire',
  heatwave: 'heatwave',
  other: 'other',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  critical: 'critical',
  warning: 'warning',
  advisory: 'advisory',
} as const;

export type AlertStatus = typeof AlertStatus[keyof typeof AlertStatus];


export const AlertStatus = {
  draft: 'draft',
  active: 'active',
  expired: 'expired',
} as const;

export type AlertChannelsItem = typeof AlertChannelsItem[keyof typeof AlertChannelsItem];


export const AlertChannelsItem = {
  sms: 'sms',
  push: 'push',
  web: 'web',
  siren: 'siren',
  radio: 'radio',
} as const;

export type ChannelDeliveryChannel = typeof ChannelDeliveryChannel[keyof typeof ChannelDeliveryChannel];


export const ChannelDeliveryChannel = {
  sms: 'sms',
  push: 'push',
  web: 'web',
  siren: 'siren',
  radio: 'radio',
} as const;

export type ChannelDeliveryStatus = typeof ChannelDeliveryStatus[keyof typeof ChannelDeliveryStatus];


export const ChannelDeliveryStatus = {
  sent: 'sent',
  sending: 'sending',
  queued: 'queued',
  unavailable: 'unavailable',
} as const;

export interface ChannelDelivery {
  channel: ChannelDeliveryChannel;
  delivered: number;
  total: number;
  status: ChannelDeliveryStatus;
}

export interface DeliveryMetrics {
  audience: number;
  delivered: number;
  acknowledged: number;
  failed: number;
  lastUpdated: string;
  channels: ChannelDelivery[];
}

export interface Alert {
  id: string;
  title: string;
  message: string;
  type: AlertType;
  severity: AlertSeverity;
  status: AlertStatus;
  location: string;
  affectedAreas: string[];
  instructions: string[];
  channels: AlertChannelsItem[];
  createdAt: string;
  /** @nullable */
  publishedAt: string | null;
  expiresAt: string;
  source: string;
  delivery: DeliveryMetrics;
}

export type AlertInputType = typeof AlertInputType[keyof typeof AlertInputType];


export const AlertInputType = {
  flood: 'flood',
  cyclone: 'cyclone',
  earthquake: 'earthquake',
  wildfire: 'wildfire',
  heatwave: 'heatwave',
  other: 'other',
} as const;

export type AlertInputSeverity = typeof AlertInputSeverity[keyof typeof AlertInputSeverity];


export const AlertInputSeverity = {
  critical: 'critical',
  warning: 'warning',
  advisory: 'advisory',
} as const;

export type AlertInputChannelsItem = typeof AlertInputChannelsItem[keyof typeof AlertInputChannelsItem];


export const AlertInputChannelsItem = {
  sms: 'sms',
  push: 'push',
  web: 'web',
  siren: 'siren',
  radio: 'radio',
} as const;

export interface AlertInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  message: string;
  type: AlertInputType;
  severity: AlertInputSeverity;
  location: string;
  affectedAreas: string[];
  instructions: string[];
  channels: AlertInputChannelsItem[];
  expiresAt: string;
  source: string;
}

export type AlertUpdateSeverity = typeof AlertUpdateSeverity[keyof typeof AlertUpdateSeverity];


export const AlertUpdateSeverity = {
  critical: 'critical',
  warning: 'warning',
  advisory: 'advisory',
} as const;

export type AlertUpdateChannelsItem = typeof AlertUpdateChannelsItem[keyof typeof AlertUpdateChannelsItem];


export const AlertUpdateChannelsItem = {
  sms: 'sms',
  push: 'push',
  web: 'web',
  siren: 'siren',
  radio: 'radio',
} as const;

export interface AlertUpdate {
  /** @minLength 3 */
  title?: string;
  /** @minLength 10 */
  message?: string;
  severity?: AlertUpdateSeverity;
  location?: string;
  affectedAreas?: string[];
  instructions?: string[];
  channels?: AlertUpdateChannelsItem[];
  expiresAt?: string;
}

export type ActivityEventSeverity = typeof ActivityEventSeverity[keyof typeof ActivityEventSeverity];


export const ActivityEventSeverity = {
  neutral: 'neutral',
  success: 'success',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface ActivityEvent {
  id: string;
  action: string;
  detail: string;
  time: string;
  severity: ActivityEventSeverity;
}

export interface DashboardSummary {
  activeAlerts: number;
  peopleReached: number;
  deliveryRate: number;
  acknowledgements: number;
  channelHealth: ChannelDelivery[];
  recentActivity: ActivityEvent[];
}

export type ListAlertsParams = {
status?: ListAlertsStatus;
audience?: string;
search?: string;
};

export type ListAlertsStatus = typeof ListAlertsStatus[keyof typeof ListAlertsStatus];


export const ListAlertsStatus = {
  draft: 'draft',
  active: 'active',
  expired: 'expired',
  all: 'all',
} as const;