# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: alerts
    description: Disaster alert operations
  - name: dashboard
    description: Authority dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List disaster alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, expired, all]
            default: all
        - name: audience
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a disaster alert draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get a disaster alert
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update a disaster alert
      parameters:
        - $ref: "#/components/parameters/AlertId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/publish:
    post:
      operationId: publishAlert
      tags: [alerts]
      summary: Publish an alert to its selected channels
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Published alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/delivery:
    get:
      operationId: getAlertDelivery
      tags: [alerts]
      summary: Get delivery metrics for an alert
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Delivery metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryMetrics"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get authority dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  parameters:
    AlertId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Alert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [flood, cyclone, earthquake, wildfire, heatwave, other]
        severity:
          type: string
          enum: [critical, warning, advisory]
        status:
          type: string
          enum: [draft, active, expired]
        location:
          type: string
        affectedAreas:
          type: array
          items:
            type: string
        instructions:
          type: array
          items:
            type: string
        channels:
          type: array
          items:
            type: string
            enum: [sms, push, web, siren, radio]
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: string
          format: date-time
        source:
          type: string
        delivery:
          $ref: "#/components/schemas/DeliveryMetrics"
      required:
        - id
        - title
        - message
        - type
        - severity
        - status
        - location
        - affectedAreas
        - instructions
        - channels
        - createdAt
        - publishedAt
        - expiresAt
        - source
        - delivery
    AlertInput:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        message:
          type: string
          minLength: 10
        type:
          type: string
          enum: [flood, cyclone, earthquake, wildfire, heatwave, other]
        severity:
          type: string
          enum: [critical, warning, advisory]
        location:
          type: string
        affectedAreas:
          type: array
          items:
            type: string
        instructions:
          type: array
          items:
            type: string
        channels:
          type: array
          items:
            type: string
            enum: [sms, push, web, siren, radio]
        expiresAt:
          type: string
          format: date-time
        source:
          type: string
      required:
        - title
        - message
        - type
        - severity
        - location
        - affectedAreas
        - instructions
        - channels
        - expiresAt
        - source
    AlertUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        message:
          type: string
          minLength: 10
        severity:
          type: string
          enum: [critical, warning, advisory]
        location:
          type: string
        affectedAreas:
          type: array
          items:
            type: string
        instructions:
          type: array
          items:
            type: string
        channels:
          type: array
          items:
            type: string
            enum: [sms, push, web, siren, radio]
        expiresAt:
          type: string
          format: date-time
      minProperties: 1
    DeliveryMetrics:
      type: object
      properties:
        audience:
          type: integer
        delivered:
          type: integer
        acknowledged:
          type: integer
        failed:
          type: integer
        lastUpdated:
          type: string
          format: date-time
        channels:
          type: array
          items:
            $ref: "#/components/schemas/ChannelDelivery"
      required:
        - audience
        - delivered
        - acknowledged
        - failed
        - lastUpdated
        - channels
    ChannelDelivery:
      type: object
      properties:
        channel:
          type: string
          enum: [sms, push, web, siren, radio]
        delivered:
          type: integer
        total:
          type: integer
        status:
          type: string
          enum: [sent, sending, queued, unavailable]
      required:
        - channel
        - delivered
        - total
        - status
    DashboardSummary:
      type: object
      properties:
        activeAlerts:
          type: integer
        peopleReached:
          type: integer
        deliveryRate:
          type: number
        acknowledgements:
          type: integer
        channelHealth:
          type: array
          items:
            $ref: "#/components/schemas/ChannelDelivery"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEvent"
      required:
        - activeAlerts
        - peopleReached
        - deliveryRate
        - acknowledgements
        - channelHealth
        - recentActivity
    ActivityEvent:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        detail:
          type: string
        time:
          type: string
          format: date-time
        severity:
          type: string
          enum: [neutral, success, warning, critical]
      required:
        - id
        - action
        - detail
        - time
        - severity