# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventSeverity } from './activityEventSeverity';

export interface ActivityEvent {
  id: string;
  action: string;
  detail: string;
  time: Date;
  severity: ActivityEventSeverity;
}