# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlertChannelsItem } from './alertChannelsItem';
import type { AlertSeverity } from './alertSeverity';
import type { AlertStatus } from './alertStatus';
import type { AlertType } from './alertType';
import type { DeliveryMetrics } from './deliveryMetrics';

export interface Alert {
  id: string;
  title: string;
  message: string;
  type: AlertType;
  severity: AlertSeverity;
  status: AlertStatus;
  location: string;
  affectedAreas: string[];
  instructions: string[];
  channels: AlertChannelsItem[];
  createdAt: Date;
  /** @nullable */
  publishedAt: Date | null;
  expiresAt: Date;
  source: string;
  delivery: DeliveryMetrics;
}