# lib/api-zod/src/generated/types/channelDelivery.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelDeliveryChannel } from './channelDeliveryChannel';
import type { ChannelDeliveryStatus } from './channelDeliveryStatus';

export interface ChannelDelivery {
  channel: ChannelDeliveryChannel;
  delivered: number;
  total: number;
  status: ChannelDeliveryStatus;
}