# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from './activityEvent';
import type { ChannelDelivery } from './channelDelivery';

export interface DashboardSummary {
  activeAlerts: number;
  peopleReached: number;
  deliveryRate: number;
  acknowledgements: number;
  channelHealth: ChannelDelivery[];
  recentActivity: ActivityEvent[];
}