# lib/api-zod/src/generated/types/deliveryMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelDelivery } from './channelDelivery';

export interface DeliveryMetrics {
  audience: number;
  delivered: number;
  acknowledged: number;
  failed: number;
  lastUpdated: Date;
  channels: ChannelDelivery[];
}